+++ src/core/plate_recognize.cpp
@@ -44,7 +44,7 @@ int CPlateRecognize::plateRecognize(Mat src, vector<string>& licenseVec)
 	int resultPD = plateDetect(src, plateVec);
 	if (resultPD == 0)
 	{
-		int num = plateVec.size();
+		int num = (int)plateVec.size();
 
 		int resultCR = 0;
 		for (int j = 0; j < num; j++)
